CANS-736: minor test cleanup
