+++ src/test/java/gov/ca/cwds/cans/rest/resource/SupervisorAuthorizationResourceTest.java
@@ -1,14 +1,10 @@
 package gov.ca.cwds.cans.rest.resource;
 
-import static gov.ca.cwds.cans.Constants.API.ASSESSMENTS;
-
 import gov.ca.cwds.cans.domain.dto.assessment.AssessmentDto;
 import java.io.IOException;
 import java.util.Stack;
 import javafx.util.Pair;
-import javax.ws.rs.core.MediaType;
 import org.apache.http.HttpStatus;
-import org.junit.After;
 import org.junit.Test;
 
 public class SupervisorAuthorizationResourceTest extends AbstractFunctionalTest {
@@ -22,18 +18,6 @@ public class SupervisorAuthorizationResourceTest extends AbstractFunctionalTest 
   public static final String FIXTURES_SUPERVISOR_00j_MERCED =
       "fixtures/supervisor/00j-merced-supervisor.json";
   private final Stack<Pair<Long, String>> cleanUpAssessmentsToUserFixtures = new Stack<>();
-
-  @After
-  public void tearDown() throws IOException {
-    while (!cleanUpAssessmentsToUserFixtures.empty()) {
-      final Pair<Long, String> assessmentIdToUserFixture = cleanUpAssessmentsToUserFixtures.pop();
-      clientTestRule
-          .withSecurityToken(assessmentIdToUserFixture.getValue())
-          .target(ASSESSMENTS + SLASH + assessmentIdToUserFixture.getKey())
-          .request(MediaType.APPLICATION_JSON_TYPE)
-          .delete();
-    }
-  }
 
   @Test
   public void getClient_success_whenAssignedToSubordinateThroughCase() throws IOException {
